remove compilator ambiguity

+++ test/net/azib/ipscan/util/SequenceIteratorTest.java
@@ -24,7 +24,7 @@ public class SequenceIteratorTest {
 	public void singleIterator() throws Exception {
 		Iterator<Integer> i = new SequenceIterator<Integer>(Arrays.asList(1).iterator());
 		assertTrue(i.hasNext());
-		assertEquals(1, i.next());
+		assertEquals(1, (int)i.next());
 		assertFalse(i.hasNext());
 	}
 
@@ -32,11 +32,11 @@ public class SequenceIteratorTest {
 	public void twoIterators() throws Exception {
 		Iterator<Integer> i = new SequenceIterator<Integer>(Arrays.asList(1, 2).iterator(), Arrays.asList(3).iterator());
 		assertTrue(i.hasNext());
-		assertEquals(1, i.next());
+		assertEquals(1, (int)i.next());
 		assertTrue(i.hasNext());
-		assertEquals(2, i.next());
+		assertEquals(2, (int)i.next());
 		assertTrue(i.hasNext());
-		assertEquals(3, i.next());
+		assertEquals(3, (int)i.next());
 		assertFalse(i.hasNext());
 	}
 
@@ -44,7 +44,7 @@ public class SequenceIteratorTest {
 	public void firstEmpty() throws Exception {
 		Iterator<Integer> i = new SequenceIterator<Integer>(Arrays.<Integer>asList().iterator(), Arrays.asList(3).iterator());
 		assertTrue(i.hasNext());
-		assertEquals(3, i.next());
+		assertEquals(3, (int)i.next());
 		assertFalse(i.hasNext());
 	}
 }
